Refactor : 클래스 명 수정
- 자잘한 클래스 명을 수정합니다.
