+++ src/main/java/nbbang/com/nbbang/global/socket/SocketAPI.java
@@ -14,7 +14,7 @@ import nbbang.com.nbbang.domain.chat.dto.ChatReadSocketDto;
 import nbbang.com.nbbang.domain.chat.dto.message.ChatAlarmResponseDto;
 import nbbang.com.nbbang.domain.chat.dto.message.ChatSendRequestDto;
 import nbbang.com.nbbang.domain.chat.dto.message.ChatSendResponseDto;
-import nbbang.com.nbbang.domain.chat.service.MessageServiceImpl;
+import nbbang.com.nbbang.domain.chat.service.MessageService;
 import nbbang.com.nbbang.domain.party.domain.Party;
 import nbbang.com.nbbang.domain.party.service.PartyService;
 import nbbang.com.nbbang.global.interceptor.CurrentMember;
@@ -32,7 +32,7 @@ import static nbbang.com.nbbang.global.socket.SocketDestination.*;
 @RequestMapping("socket/develop")
 public class SocketAPI {
     @Autowired
-    MessageServiceImpl messageServiceImpl;
+    MessageService messageService;
     @Autowired
     CurrentMember currentMember;
     @Autowired
@@ -42,7 +42,7 @@ public class SocketAPI {
     @ApiResponse(responseCode = "200", description = "OK", content = @Content(mediaType = "application/json", schema = @Schema(implementation = ChatSendResponseDto.class)))
     @PostMapping("/chatting")
     public SocketSendDto socketEmitApi(@RequestBody ChatSendRequestDto chatSendRequestDto){
-        Message message = messageServiceImpl.send(1L, 1L, chatSendRequestDto.getContent());
+        Message message = messageService.send(1L, 1L, chatSendRequestDto.getContent());
         ChatSendResponseDto data = ChatSendResponseDto.createByMessage(message, currentMember.id());
         return SocketSendDto.builder().type(CHATTING).data(data).build();
     }
@@ -70,7 +70,7 @@ public class SocketAPI {
     @ApiResponse(responseCode = "200", description = "OK", content = @Content(mediaType = "application/json", schema = @Schema(implementation = ChatAlarmResponseDto.class)))
     @PostMapping("/global")
     public SocketSendDto global(){
-        Message message = messageServiceImpl.send(1L, 2L, "다른 사람이 보낸 채팅");
+        Message message = messageService.send(1L, 2L, "다른 사람이 보낸 채팅");
         ChatSendResponseDto data = ChatSendResponseDto.createByMessage(message, currentMember.id());
         return SocketSendDto.builder().type(GLOBAL).data(ChatAlarmResponseDto.create(Party.builder().id(0L).title("test party").build(), data)).build();
     }
